put code in search/views.py instead of search/__init__.py
